requirements.txt
boto==2.27.0
python_gnupg==0.3.6